create landing page
Created landing page with following content
Header with logo on the left and list of links
displayd flex with space-between to kepp items on
the edges
Hero with image on the right and content on the
left, Heding paragraph and button
about with flex grid column inside 4 cards each
has an image and descriprion
Testimonial with a quote and capture on the right
display flex to keep sign on the right
CTA section with a banner inside h2 heading and a
text on the left and button on the right. Also
display flex and space between to keep items on
the edges
footer
with a text inside. text aligned center for
simplicity